fix(build): more sign-conversions and the third bench-harness clone
Three remaining strict-Linux-compile errors closed.

1) tests/ipc/test_sprint5_gate_concurrency.cpp
   - Producer-loop `i` captured into the lambda was `int`; both
     `sendGate[i]` and (in the verification loop) `maxObserved[i]`
     are std::vector subscripts that want size_t. Wrapped each
     subscript site in `static_cast<std::size_t>(...)`; the producer
     lambda also gets a local `const auto ui = static_cast<std::size_t>(i)`
     to keep the body readable.

2) tests/ipc/bench_ipc_with_gate.cpp:110
   - Third copy of the `pct * v.size()` percentile lambda. Same fix
     as bench_harness.h and tests/core/bench_validate.cpp: explicit
     static_cast<double>(v.size()).

3) tests/ipc/bench_latency.cpp:121
   - CPU_SET(iCore, ...) sign-conversion. Same fix as bench_throughput
     and bench_throughput_mpsc: static_cast<std::size_t>(iCore), plus
     an `if (iCore < 0) return;` guard so the cast is provably safe.

Pre-emptive grep sweep confirms zero remaining sites of either pattern
in src/ or tests/.

After pull:
  cd ~/projects/Astra
  git pull
  ./scripts/compile-astra.sh

+++ tests/ipc/bench_ipc_with_gate.cpp
@@ -107,8 +107,13 @@ static Stats fnSummarise(std::vector<uint64_t>& v, double fTscPerNs)
     if (v.empty()) return s;
     std::sort(v.begin(), v.end());
     auto fnAt = [&](double pct) {
-        size_t i = std::min<size_t>(static_cast<size_t>(pct * v.size()),
-                                    v.size() - 1);
+        // Explicit double-cast on v.size() — same fix as bench_harness.h
+        // and tests/core/bench_validate.cpp. Three separate copies of
+        // this percentile lambda exist across the codebase; this is the
+        // third site to receive it.
+        size_t i = std::min<size_t>(
+            static_cast<size_t>(pct * static_cast<double>(v.size())),
+            v.size() - 1);
         return static_cast<double>(v[i]) / fTscPerNs;
     };
     long double sum = 0;
